Will read in data, still doesn't log it.
Lots of bug fixes from previous version.  For some reason the code I put in to disable interrupts if(Serial1.available()) and else reenable them was causing the IMU to not ever be read, probably because interrupts weren't getting reenabled for some reason.  I got rid of those lines, and GPS data still read in fine as far as I could tell... maybe interrupts only mess up software serial.

Also, major note about the GPS:  When using hardware serial, there is a mistake in the adafruit code. See: https://forums.adafruit.com/viewtopic.php?f=22&t=74869&p=379587
GPS object needs to be initialized with (&Serial1) not (&mySerial).

Note about using SD card: Need to either flush or close file after writing to it, because that is when the information actually gets written to the card.  I tried closing the file every time after writing to it and then reopening it again, and this seemed to work, but for some reason when I just flushed the file it doesn't work.  I would like to